scripts/monitor/pyproject.toml
[project]
name = "monitor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pyelftools>=0.32",
    "pyocd>=0.44.0",
]
